requirements.txt
tornado
requests
scikit-image
mysql-connector-python